requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.6
django-bootstrap4==23.4
django-chartjs==2.3.0
django-localflavor==4.0
idna==3.6
psycopg2==2.9.9
python-decouple==3.8
python-stdnum==1.19
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
